requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
azure-core==1.32.0
azure-storage-blob==12.24.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
distro==1.9.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isodate==0.7.2
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.19.1
networkx==3.4.2
numpy==1.26.4
openai==1.58.1
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.29.2
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.55
ultralytics-thop==2.0.13
urllib3==2.3.0
